refactor(wfs-surce): Move wfsSource to common/layers
