+++ projects/hslayers/common/layers/public-api.ts
@@ -1,3 +1,5 @@
 export * from './hs.source.SPOI';
 export * from './hs.source.SparqlJson';
 export * from './hs.source.interpolated';
+export * from './hs.source.WfsSource';
+export * from './wfs-get-feature';
